+++ lib/metro/stubs/model_stub.dart
@@ -1,29 +1,17 @@
-String modelStub({String? modelName, bool? isStorable}) => '''
-import 'package:nylo_framework/nylo_framework.dart';\n\nclass $modelName implements JsonOperations ${isStorable == true ? "extends Storable " : ""}{
+String modelStub({String? modelName}) => '''
+import 'package:nylo_framework/nylo_framework.dart';
+
+/// $modelName Model.
+
+class $modelName extends Model {
   $modelName();
   
-  @override
   $modelName.fromJson(data) {
 
   }
 
-  @override
   toJson() {
 
   }
-${isStorable == true ? '''
-    /// Example [foo] below. 
-    /// New to Storable? Learn more https://nylo.dev/docs/storage#introduction-to-storable-models
-    late String foo; 
-
-    @override
-    toStorage() => {
-      // "foo": this.foo
-    };
-
-    @override
-    fromStorage(dynamic data) {
-      // this.foo = data['foo'];
-    }
-''' : ""}}
+}
 ''';